remove ClientId from Web Api Configuration.
