+++ Okta.AspNet.Abstractions/OktaWebApiOptions.cs
@@ -3,6 +3,8 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
+using System;
+
 namespace Okta.AspNet.Abstractions
 {
     public class OktaWebApiOptions : OktaWebOptions
@@ -10,5 +12,8 @@ namespace Okta.AspNet.Abstractions
         public static readonly string DefaultAudience = "api://default";
 
         public string Audience { get; set; } = DefaultAudience;
+
+        [Obsolete("ClientId is deprecated")]
+        public string ClientId { get; set; }
     }
 }
