+++ Okta.AspNet.Abstractions/OktaWebOptionsValidator.cs
@@ -62,20 +62,6 @@ namespace Okta.AspNet.Abstractions
                 throw new ArgumentNullException(nameof(options.OktaDomain), "It looks like there's a typo in your Okta domain. You can copy your domain from the Okta Developer Console.");
             }
 
-            if (string.IsNullOrEmpty(options.ClientId))
-            {
-                throw new ArgumentNullException(
-                    nameof(options.ClientId),
-                    "Your client ID is missing. You can copy it from the Okta Developer Console in the details for the Application you created.");
-            }
-
-            if (options.ClientId.IndexOf("{ClientId}", StringComparison.OrdinalIgnoreCase) >= 0)
-            {
-                throw new ArgumentNullException(
-                    nameof(options.ClientId),
-                    "Replace {clientId} with the client ID of your Application. You can copy it from the Okta Developer Console in the details for the Application you created.");
-            }
-
             ValidateInternal((T)options);
         }
     }
