+++ Okta.AspNet/OktaMvcOptions.cs
@@ -11,6 +11,8 @@ namespace Okta.AspNet
     {
         public string ClientSecret { get; set; }
 
+        public string ClientId { get; set; }
+
         public string RedirectUri { get; set; }
 
         public string PostLogoutRedirectUri { get; set; }
